Feat/gahee

+++ src/main/java/team/project/fiverockrun/domain/price/dto/request/SectionPriceRequest.java
@@ -13,11 +13,11 @@ public class SectionPriceRequest {
 
     @NotNull(message = "출발역의 정차 순서는 필수입니다.")
     @Min(value = 1, message = "출발 순서는 1 이상이어야 합니다.")
-    private Integer startOrder;
+    private Long startOrder;
 
     @NotNull(message = "도착역의 정차 순서는 필수입니다.")
     @Min(value = 2, message = "도착 순서는 2 이상이어야 합니다.")
-    private Integer endOrder;
+    private Long endOrder;
 
     @NotNull(message = "좌석 등급은 필수입니다.")
     private SeatType seatType;
